extract app from wasm-rpc-stubgen
